Disable EXTRACT_ALL in Doxygen
I've decided that to keep the documentation as tidy as possible, only
the necessary things should be documented, and for that reason it's not
worth it to have EXTRACT_ALL = True. Unless I think something needs
explaining in some header file, I won't annotate it and Doxygen will not
even display that file as clickable etc.